Ensure the the hostname -> ip-address DNS resolution is done every time the graphite socket is recreated

+++ src/main/java/org/jmxtrans/agent/FileOverwriterOutputWriter.java
@@ -91,7 +91,9 @@ public class FileOverwriterOutputWriter extends AbstractOutputWriter {
         } catch (IOException e) {
             // silently skip
         }
-        temporaryFile.delete();
+        if (temporaryFile != null) {
+           temporaryFile.delete();
+        }
         temporaryFile = null;
 
     }
